# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Baba Abdullah Kitchen API
  description: |
    Microservices REST API for Baba Abdullah Kitchen PWA dashboard.
    Supports product management, inventory, orders, sales reports, and promotions.
  version: 1.0.0
  contact:
    name: Baba Abdullah Kitchen
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://baba-menu.umsoftware.id
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        created_at: { type: string, format: date-time }
...
    Promotion:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        discount_type: { type: string, enum: [percent, fixed] }
        discount_value: { type: number }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        is_active: { type: integer, enum: [0, 1] }
        created_at: { type: string, format: date-time }
        products:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
...
paths:
  /api/auth/login:
    post:
      summary: Admin login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: admin123 }
      responses:
        '200':
          description: JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  username: { type: string }
        '401':
          description: Invalid credentials
  /api/categories:
    get:
      summary: List all categories
      tags: [Categories]
      responses:
        '200':
          description: Array of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Create category
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        '201':
          description: Category created
  /api/categories/{id}:
    put:
      summary: Update category
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        '200':
          description: Category updated
    delete:
      summary: Delete category
      tags: [Categories]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
...
  /api/promotions:
    get:
      summary: List all promotions
      tags: [Promotions]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
    post:
      summary: Create promotion
      tags: [Promotions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, discount_type, discount_value, start_date, end_date]
              properties:
                name: { type: string }
                description: { type: string }
                discount_type: { type: string, enum: [percent, fixed] }
                discount_value: { type: number }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                is_active: { type: integer }
                product_ids:
                  type: array
                  items: { type: integer }
      responses:
        '201':
          description: Promotion created
  /api/promotions/{id}:
    get:
      summary: Get promotion
      tags: [Promotions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Promotion detail
    put:
      summary: Update promotion
      tags: [Promotions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                discount_type: { type: string, enum: [percent, fixed] }
                discount_value: { type: number }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                is_active: { type: integer }
                product_ids:
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete promotion
      tags: [Promotions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
  /api/promotions/{id}/products:
    post:
      summary: Attach/detach products to promotion
      tags: [Promotions]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_ids:
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: Products updated